requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cssselect==1.3.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.2
hf-xet==1.1.0
huggingface-hub==0.31.1
idna==3.10
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.0
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pillow==11.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
readability-lxml==0.8.4.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
safetensors==0.5.3
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
sympy==1.14.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
